+++ src/main/java/org/apache/commons/lang3/time/FastDatePrinter.java
@@ -409,7 +409,7 @@ public class FastDatePrinter implements DatePrinter, Serializable {
         } else if (obj instanceof Calendar) {
             return format((Calendar) obj, toAppendTo);
         } else if (obj instanceof Long) {
-            return format((Long) obj, toAppendTo);
+            return format(((Long) obj).longValue(), toAppendTo);
         } else {
             throw new IllegalArgumentException("Unknown class: " +
                 (obj == null ? "<null>" : obj.getClass().getName()));
